@@ -7,7 +7,7 @@
     "start": "ts-node-dev ./src/server.ts"
   },
   "engines": {
-    "node": "12.18.0",
+    "node": "14.9.0",
     "yarn": "1.22.4"
   },
   "keywords": [],
@@ -15,6 +15,7 @@
   "license": "MIT",
   "dependencies": {
     "@types/bcryptjs": "^2.4.2",
+    "@types/cors": "^2.8.7",
     "@types/express": "4.17.0",
     "@types/jest": "^26.0.4",
     "@types/jsonwebtoken": "^8.5.0",
